Refactor package name
