+++ src/main/java/de/unistuttgart/iste/gits/user_service/config/package-info.java
@@ -1,4 +1,4 @@
 /**
  * This package should contain any classes that are used to configure the application.
  */
-package de.unistuttgart.iste.gits.membership.config;
+package de.unistuttgart.iste.gits.user_service.config;
