+++ src/main/java/de/unistuttgart/iste/gits/user_service/dapr/package-info.java
@@ -1,4 +1,4 @@
 /**
  * This package should contain all classes that are used to communicate with Dapr, e.g. using pub sub.
  */
-package de.unistuttgart.iste.gits.membership.dapr;
+package de.unistuttgart.iste.gits.user_service.dapr;
