+++ src/main/java/de/unistuttgart/iste/gits/user_service/exception/package-info.java
@@ -3,4 +3,4 @@
  * More info about exception handling with GraphQL can be found here:
  * https://www.baeldung.com/spring-graphql-error-handling
  */
-package de.unistuttgart.iste.gits.membership.exception;
+package de.unistuttgart.iste.gits.user_service.exception;
