+++ src/main/java/de/unistuttgart/iste/gits/user_service/persistence/dao/package-info.java
@@ -10,4 +10,4 @@
  * To avoid having to implement getters, setters, builders, etc. we use <a href=https://projectlombok.org/>Lombok</a>.
  * The Lombok annotations (e.g. {@code @Data}) should be used in the entities.
  */
-package de.unistuttgart.iste.gits.membership.persistence.dao;
+package de.unistuttgart.iste.gits.user_service.persistence.dao;
